move types from various locations to types.ts file :ambulance:
