+++ library/src/helpers/index.ts
@@ -1,3 +1,2 @@
 export * from './parser';
 export * from './beautifier';
-export * from './fetchSchema';
