+++ library/src/index.ts
@@ -1,8 +1,10 @@
 import AsyncApi from './containers/AsyncApi/AsyncApi';
-export { AsyncApiProps } from './containers/AsyncApi/AsyncApi';
 export { ThemeInterface } from './theme/theme';
 export { ConfigInterface } from './config/config';
-export { FetchingSchemaInterface } from './helpers';
-export { AsyncApi as AsyncApiInterface } from './types';
+export {
+  AsyncApi as AsyncApiInterface,
+  AsyncApiProps,
+  FetchingSchemaInterface,
+} from './types';
 
 export default AsyncApi;
